+++ kraftwerk-core/src/main/java/fr/insee/kraftwerk/core/metadata/LunaticReader.java
@@ -50,8 +50,8 @@ public class LunaticReader {
 			JsonNode variablesNode = rootNode.get(VARIABLES);
 			variablesNode.forEach(variableNode -> {
 				if (variableNode.get("variableType").asText().equals("CALCULATED")) {
-					String formula = isLunaticV2 ? variableNode.get("expression").get(VALUE).asText()
-							: variableNode.get("expression").asText();
+					String formula = isLunaticV2 ? variableNode.get("expression").get(VALUE).asText().replace("\r\n","")
+							: variableNode.get("expression").asText().replace("\r\n","");
 					CalculatedVariable calculatedVariable = new CalculatedVariable(variableNode.get("name").asText(),
 							formula);
 					JsonNode dependantVariablesNode = variableNode.get(BINDING_DEPENDENCIES);
@@ -135,6 +135,7 @@ public class LunaticReader {
 	 * @return The variables found in the Lunatic specification.
 	 */
 	public static MetadataModel getMetadataFromLunatic(Path lunaticFile, FileUtilsInterface fileUtilsInterface) {
+		log.info("Mode Lunatic-only : start reading file {} ",lunaticFile.getFileName());
 		JsonNode rootNode;
 		try {
 			rootNode = JsonReader.read(lunaticFile, fileUtilsInterface);